fix: resolve CI failures, security hotspots, and code duplication
- Centralize IP extraction logic in src/utils/ip-utils.ts.
- Use validator.isIP for secure extraction in rate-limit and logger modules.
- Pin google-github-actions/run-gemini-cli to @v0 in workflows.
- Update unit tests to verify centralized IP extraction.
- Maintain 94% coverage for rate-limit utility.
